Part 9.4
Fixed:
Npcs rendering above buildings
Particles not showing
New bugs found:
Demons/shooters not attacking unless close range
Names not centered
Demon fireball not  centered when shooting
Player no clip monsters (revert this back)

+++ src/game/entities/projectiles/Arrow.java
@@ -33,7 +33,7 @@ public class Arrow extends Projectile {
 	 * @param damage
 	 *            the damage of this arrow
 	 */
-	public Arrow(Level level, double x, double y, int xPos, int yPos, Mob mob, int damage) {
+	public Arrow(Level level, int x, int y, int xPos, int yPos, Mob mob, int damage) {
 		super(level, x, y, 2, 1, 1, 6, xPos, yPos, mob, damage, SoundHandler.laser);
 
 		adjustOffset(mob);
@@ -55,7 +55,7 @@ public class Arrow extends Projectile {
 	 * @param damage
 	 *            the damage of this arrow
 	 */
-	public Arrow(Level level, double x, double y, Direction direction, Mob mob, int damage) {
+	public Arrow(Level level, int x, int y, Direction direction, Mob mob, int damage) {
 		super(level, x, y, 2, 1, 1, 6, direction, mob, damage, SoundHandler.laser);
 
 		adjustOffset(mob);
