+++ src/game/entities/projectiles/FireBall.java
@@ -28,7 +28,7 @@ public class FireBall extends Projectile {
 	 * @param mob the mob that is firing
 	 * @param damage the damage of this fireball
 	 */
-	public FireBall(Level level, double x, double y, int xPos, int yPos, Mob mob, int damage) {
+	public FireBall(Level level, int x, int y, int xPos, int yPos, Mob mob, int damage) {
 		super(level, x, y, SIZE, SIZE, 0, 1, xPos, yPos, mob, damage, SoundHandler.fireball);
 	}
 
@@ -41,7 +41,7 @@ public class FireBall extends Projectile {
 	 * @param mob the mob that is firing
 	 * @param damage the damage of this fireball
 	 */
-	public FireBall(Level level, double x, double y, Direction direction, Mob mob, int damage) {
+	public FireBall(Level level, int x, int y, Direction direction, Mob mob, int damage) {
 		super(level, x, y, SIZE, SIZE, 0, 1, direction, mob, damage, SoundHandler.fireball);
 	}
