+++ src/game/entities/projectiles/Laser.java
@@ -33,7 +33,7 @@ public class Laser extends Projectile {
 	 * @param damage
 	 *            the damage of this laser
 	 */
-	public Laser(Level level, double x, double y, int xPos, int yPos, Mob mob, int damage) {
+	public Laser(Level level, int x, int y, int xPos, int yPos, Mob mob, int damage) {
 		super(level, x, y, 2, 1, 1, 6, xPos, yPos, mob, damage, SoundHandler.laser);
 	}
 
@@ -53,7 +53,7 @@ public class Laser extends Projectile {
 	 * @param damage
 	 *            the damage of this laser
 	 */
-	public Laser(Level level, double x, double y, Direction direction, Mob mob, int damage) {
+	public Laser(Level level, int x, int y, Direction direction, Mob mob, int damage) {
 		super(level, x, y, 2, 1, 1, 6, direction, mob, damage, SoundHandler.laser);
 	}
